gen3release-sdk/pyproject.toml
[tool.poetry]
name = "gen3release"
version = "0.1.8"
description = "Gen3 Release Management SDK"
authors = ["CTDS UChicago <cdis@uchicago.edu>"]
packages = [
    { include = "gen3release/config" },
    { include = "gen3release/filesys" },
    { include = "gen3release/gith" },
    { include = "gen3release/env_cli.py" }
]

[tool.poetry.dependencies]
python = "^3.6"
PyGitHub = "^1.47"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

[tool.poetry.scripts]
gen3release = 'gen3release.env_cli:main'
